// src/app/models/listing.ts
export interface listing {
    id: number
    type: string
    status: string
    numberOfRooms: number
    numberOfBathrooms: number
    numberOfKitchens: number
    price: number
    location: string
    area: number
    parking: number
    description: string
    mainImageFileName: string
    otherImagesFileNames: string
    agentId: number
}